### 2023-07-10 #5640 @editorialbot
[PRE REVIEW]: BayesFlow: Amortized Bayesian Workflows With Neural Networks
**Submitting author:** @marvinschmitt (<a href="http://orcid.org/0000-0002-7967-4723">Marvin Schmitt</a>)
**Repository:** https://github.com/stefanradev93/BayesFlow
**Branch with paper.md** (empty if default branch): joss-submission
**Version:** v1.1.1
