### 2023-07-10 #5640 @Kevin-Mattheus-Moerman
@marvinschmitt thanks for this submission. I am one of the AEiCs for JOSS and here to help process initial steps. I have started looking for a handling editor.
For the moment, can you please see if you can trim down your .bib file to only include the references you use (perhaps you do use all, in which case no change is needed). In addition, can you check those potentially missing/invalid DOIs :point_up: ? FYI you can use `@editorialbot check references` to check those DOIs again, and you can use `@editorialbot generate pdf` to update the draft.
