### 2023-07-11 19aac4e
trim bibliography
