+++ JOSS/paper.bib
@@ -1,1344 +1,356 @@
-@article{brandl2010sphinx,
-    year = {2010},
-    title = {Sphinx documentation},
-    journal = {URL http://sphinx-doc. org/sphinx. pdf},
-    author = {Brandl, Georg},
-}
-@book{aggarwal2018neural,
-    year = {2018},
-    title = {Neural networks and deep learning},
-    publisher = {Springer},
-    author = {Aggarwal, Charu C and others},
-}
-@article{brms1,
-    year = {2017},
-    pages = {1--28},
-    title = {{brms}: An {R} Package for {B}ayesian Multilevel Models using {Stan}},
-    number = {1},
-    volume = {80},
-    journal = {Journal of Statistical Software},
-    encoding = {UTF-8},
-    author = {B\"{u}rkner, Paul-Christian},
-}
-@article{brms2,
-    year = {2018},
-    pages = {395--411},
-    title = {Advanced {B}ayesian Multilevel Modeling with the {R} Package {brms}},
-    number = {1},
-    volume = {10},
-    journal = {The {R} Journal},
-    encoding = {UTF-8},
-    author = {B\"{u}rkner, Paul-Christian},
-}
-@article{brms3,
-    year = {2020},
-    pages = {1--54},
-    title = {{B}ayesian Item Response Modelling in {R} with {brms} and {Stan}},
-    journal = {Journal of Statistical Software},
-    author = {B\"{u}rkner, Paul-Christian},
-}
-@article{buerkner2020lfo,
-    year = {2020},
-    pages = {2499--2523},
-    title = {Approximate leave-future-out cross-validation for {B}ayesian time series models},
-    number = {14},
-    volume = {90},
-    journal = {Journal of Statistical Computation and Simulation},
-    author = {B\"{u}rkner, Paul-Christian and Gabry, Jonah and Vehtari, Aki},
-}
-@article{byrne2001act,
-    year = {2001},
-    pages = {41--84},
-    title = {{ACT-R/PM} and menu selection: Applying a cognitive architecture to {HCI}},
-    number = {1},
-    volume = {55},
-    journal = {International Journal of Human-Computer Studies},
-    publisher = {Elsevier},
-    author = {Byrne, Michael D},
-}
-@article{hermans2019likelihood,
-    year = {2019},
-    title = {Likelihood-free {MCMC} with approximate likelihood ratios},
-    journal = {arXiv preprint},
-    author = {Hermans, Joeri and Begy, Volodimir and Louppe, Gilles},
-}
-@article{wieschen2020jumping,
-    year = {2020},
-    pages = {120--132},
-    title = {Jumping to conclusion? a {L}{\'e}vy flight model of decision making},
-    number = {2},
-    volume = {16},
-    journal = {The Quantitative Methods for Psychology},
-    author = {Wieschen, Eva Marie and Voss, Andreas and Radev, Stefan},
-}
-@article{akeret2015,
-    year = {2015},
-    pages = {1--20},
-    title = {Approximate {B}ayesian computation for forward modeling in cosmology},
-    number = {08},
-    volume = {2015},
-    journal = {Journal of Cosmology and Astroparticle Physics},
-    publisher = {IOP Publishing},
-    author = {Akeret, Jo{\"e}l and Refregier, Alexandre and Amara, Adam and Seehars, Sebastian and Hasner, Caspar},
-}
-@article{anderson1996act,
-    year = {1996},
-    pages = {355--365},
-    title = {{ACT}: A simple theory of complex cognition},
-    number = {4},
-    volume = {51},
-    journal = {American Psychologist},
-    publisher = {American Psychological Association (PsycARTICLES)},
-    author = {Anderson, John Robert},
-}
-@article{anderson1997act,
[…]
