### 2023-07-11 fbed04d
add missing DOIs
Addresses missing DOIs in https://github.com/openjournals/joss-reviews/issues/5640#issuecomment-1630400099
