+++ JOSS/paper.bib
@@ -16,6 +16,7 @@
   NUMBER = {6},
   PAGES = {126},
   TITLE = {Measuring QCD splittings with invertible networks},
+  doi = {10.21468/SciPostPhys.10.6.126},
   VOLUME = {10},
 }
 
@@ -82,6 +83,7 @@
   PUBLISHER = {PsyArXiv},
   DATE = {2022},
   TITLE = {A general integrative neurocognitive modeling framework to jointly describe EEG and decision-making on single trials},
+  DOI = {10.1007/s42113-023-00167-4}
 }
 
 @ARTICLE{greenberg2019automatic,
@@ -102,6 +104,7 @@
   PAGES = {5408},
   TITLE = {Towards reliable parameter extraction in MEMS final module testing using bayesian inference},
   VOLUME = {22},
+  DOI = {10.3390/s22145408}
 }
 
 @ARTICLE{lintusaari2018elfi,
@@ -167,6 +170,7 @@
   DATE = {2020},
   JOURNALTITLE = {arXiv preprint},
   TITLE = {Amortized {B}ayesian model comparison with evidential deep learning},
+  DOI = {10.1109/TNNLS.2021.3124052}
 }
 
 @ARTICLE{radev2021outbreakflow,
@@ -178,6 +182,7 @@
   PAGES = {e1009472},
   TITLE = {OutbreakFlow: Model-based Bayesian inference of disease outbreak dynamics with invertible neural networks and its application to the COVID-19 pandemics in Germany},
   VOLUME = {17},
+  DOI = {10.1371/journal.pcbi.1009472}
 }
 
 @ARTICLE{radev2020bayesflow,
@@ -204,6 +209,7 @@
   PAGES = {32},
   TITLE = {Graphical test for discrete uniformity and its applications in goodness-of-fit evaluation and multiple sample comparison},
   VOLUME = {32},
+  DOI = {doi.org/10.1007/s11222-022-10090-6}
 }
 
 @ARTICLE{Salvatier2016,
@@ -265,6 +271,7 @@
   PAGES = {104082},
   TITLE = {Estimation of agent-based models using Bayesian deep learning approach of BayesFlow},
   VOLUME = {125},
+  DOI = {10.1016/j.jedc.2021.104082}
 }
 
 @ARTICLE{siahkoohi2023reliable,
@@ -276,6 +283,7 @@
   PAGES = {R297--R322},
   TITLE = {Reliable amortized variational inference with physics-based latent distribution correction},
   VOLUME = {88},
+  DOI = {doi.org/10.1190/geo2022-0472.1}
 }
 
 @ARTICLE{sokratous2023ask,
@@ -286,6 +294,7 @@
   PAGES = {100418},
   TITLE = {How to ask twenty questions and win: Machine learning tools for assessing preferences from small samples of willingness-to-pay prices},
   VOLUME = {48},
+  doi = {doi.org/10.1016/j.jocm.2023.100418}
 }
 
 @ARTICLE{talts2018,
@@ -301,6 +310,7 @@
   JOURNALTITLE = {AIAA Scitech 2022 Forum},
   PAGES = {0631},
   TITLE = {Inverse design under uncertainty using conditional Normalizing Flows},
+  DOI = {10.2514/6.2022-0631}
 }
 
 @ARTICLE{vaswani2017attention,
@@ -320,6 +330,7 @@
   PAGES = {055311},
   TITLE = {Variational inference of fractional Brownian motion with linear computational complexity},
   VOLUME = {106},
+  DOI = {PhysRevE.106.055311}
 }
 
 @ARTICLE{von2022mental,
@@ -331,6 +342,7 @@
   PAGES = {700--708},
[…]
