### 2023-07-11 45389ad
fix doi.org/ prefixed
addresses https://github.com/openjournals/joss-reviews/issues/5640#issuecomment-1630422769
