+++ JOSS/paper.bib
@@ -33,15 +33,13 @@
 @ARTICLE{burkner_models_2022,
   ABSTRACT = {Probabilistic (Bayesian) modeling has experienced a surge of applications in almost all quantitative sciences and industrial areas. This development is driven by a combination of several factors, including better probabilistic estimation algorithms, flexible software, increased computing power, and a growing awareness of the benefits of probabilistic learning. However, a principled Bayesian model building workflow is far from complete and many challenges remain. To aid future research and applications of a principled Bayesian workflow, we ask and provide answers for what we perceive as two fundamental questions of Bayesian modeling, namely (a) "What actually is a Bayesian model?" and (b) "What makes a good Bayesian model?". As an answer to the first question, we propose the PAD model taxonomy that defines four basic kinds of Bayesian models, each representing some combination of the assumed joint distribution of all (known or unknown) variables (P), a posterior approximator (A), and training data (D). As an answer to the second question, we propose ten utility dimensions according to which we can evaluate Bayesian models holistically, namely, (1) causal consistency, (2) parameter recoverability, (3) predictive performance, (4) fairness, (5) structural faithfulness, (6) parsimony, (7) interpretability, (8) convergence, (9) estimation speed, and (10) robustness. Further, we propose two example utility decision trees that describe hierarchies and trade-offs between utilities depending on the inferential goals that drive model building and testing.},
   AUTHOR = {Bürkner, Paul-Christian and Scholz, Maximilian and Radev, Stefan T.},
-  URL = {http://arxiv.org/abs/2209.02439},
   DATE = {2022},
   FILE = {arXiv Fulltext PDF:/Users/paul.buerkner/Zotero/storage/RXAMU46Q/B\"{u}rkner et al. - 2022 - Some models are useful, but how do we know which o.pdf:application/pdf;arXiv.org Snapshot:/Users/paul.buerkner/Zotero/storage/QIBQDVLQ/2209.html:text/html},
   JOURNALTITLE = {arXiv preprint},
   KEYWORDS = {Statistics - Methodology},
-  NOTE = {arXiv:2209.02439 [stat]},
+  NOTE = {arXiv:2209.02439},
   SHORTTITLE = {Some models are useful, but how do we know which ones?},
   TITLE = {Some models are useful, but how do we know which ones? {Towards} a unified {Bayesian} model taxonomy},
-  URLDATE = {2023-01-16},
 }
 
 @ARTICLE{d2020bayesian,
@@ -209,13 +207,12 @@
   PAGES = {32},
   TITLE = {Graphical test for discrete uniformity and its applications in goodness-of-fit evaluation and multiple sample comparison},
   VOLUME = {32},
-  DOI = {doi.org/10.1007/s11222-022-10090-6}
+  DOI = {10.1007/s11222-022-10090-6}
 }
 
 @ARTICLE{Salvatier2016,
   AUTHOR = {Salvatier, John and Wiecki, Thomas V. and Fonnesbeck, Christopher},
   PUBLISHER = {PeerJ},
-  URL = {https://doi.org/10.7717/peerj-cs.55},
   DATE = {2016},
   DOI = {10.7717/peerj-cs.55},
   JOURNALTITLE = {{PeerJ} Computer Science},
@@ -238,7 +235,6 @@
 @ARTICLE{schaelte2022pyabc,
   AUTHOR = {Schälte, Yannik and Klinger, Emmanuel and Alamoudi, Emad and Hasenauer, Jan},
   PUBLISHER = {The Open Journal},
-  URL = {https://doi.org/10.21105/joss.04304},
[…]
