### 2023-07-11 bae89e4
Update HierarchicalNetwork documentation
