+++ README.md
@@ -141,18 +141,17 @@ model parameters as a result of conditioning on the data.
 
 - Radev, S. T., Mertens, U. K., Voss, A., Ardizzone, L., & Köthe, U. (2020).
 BayesFlow: Learning complex stochastic models with invertible neural networks.
-<em>IEEE Transactions on Neural Networks and Learning Systems</em>, available
-for free at: https://arxiv.org/abs/2003.06281.
+<em>IEEE Transactions on Neural Networks and Learning Systems, 33(4)</em>, 1452-1466.
 
 - Radev, S. T., Graw, F., Chen, S., Mutters, N. T., Eichel, V. M., Bärnighausen, T., & Köthe, U. (2021).
-OutbreakFlow: Model-based Bayesian inference of disease outbreak dynamics with invertible neural networks and its application to the COVID-19 pandemics in Germany. <em>PLoS computational biology</em>, 17(10), e1009472.
+OutbreakFlow: Model-based Bayesian inference of disease outbreak dynamics with invertible neural networks and its application to the COVID-19 pandemics in Germany. <em>PLoS computational biology, 17(10)</em>, e1009472.
 
 - Bieringer, S., Butter, A., Heimel, T., Höche, S., Köthe, U., Plehn, T., & Radev, S. T. (2021).
-Measuring QCD splittings with invertible networks. <em>SciPost Physics</em>, 10(6), 126.
+Measuring QCD splittings with invertible networks. <em>SciPost Physics, 10(6)</em>, 126.
 
 - von Krause, M., Radev, S. T., & Voss, A. (2022).
 Mental speed is high until age 60 as revealed by analysis of over a million participants.
-<em>Nature Human Behaviour</em>, 6(5), 700-708.
+<em>Nature Human Behaviour, 6(5)</em>, 700-708.
 
 ## Model Misspecification
 
@@ -174,7 +173,7 @@ The amortizer knows how to combine its losses and you can inspect the summary sp
 
 ### References and Further Reading
 
-- Schmitt, M., Bürkner P. C., Köthe U., & Radev S. T. (2021). Detecting Model
+- Schmitt, M., Bürkner P. C., Köthe U., & Radev S. T. (2022). Detecting Model
 Misspecification in Amortized Bayesian Inference with Neural Networks. <em>ArXiv
 preprint</em>, available for free at: https://arxiv.org/abs/2112.08866
 
@@ -283,13 +282,13 @@ Learning both densities enables us to approximate marginal likelihoods or perfor
 
 ### References and Further Reading
 
-Radev, S. T., Schmitt, M., Pratz, V., Picchini, U., Köthe, U., & Bürkner, P. C. (2023). 
-JANA: Jointly Amortized Neural Approximation of Complex Bayesian Models. <em>arXiv preprint</em>,
-available for free at: https://arxiv.org/abs/2302.09125
+Radev, S. T., Schmitt, M., Pratz, V., Picchini, U., Köthe, U., & Bürkner, P.-C. (2023). 
+JANA: Jointly amortized neural approximation of complex Bayesian models. 
+*Proceedings of the Thirty-Ninth Conference on Uncertainty in Artificial Intelligence, 216*, 1695-1706. 
+([arXiv](https://arxiv.org/abs/2302.09125))([PLMR](https://proceedings.mlr.press/v216/radev23a.html))
 
 ## Support
[…]
