### 2023-07-16 1c8e65a
Remove old docs [skip ci]
