+++ README.md
@@ -7,13 +7,14 @@ Welcome to our BayesFlow library for efficient simulation-based Bayesian workflo
 
 For starters, check out some of our walk-through notebooks:
 
-1. [Quickstart amortized posterior estimation](docs/source/tutorial_notebooks/Intro_Amortized_Posterior_Estimation.ipynb)
-2. [Detecting model misspecification in posterior inference](docs/source/tutorial_notebooks/Model_Misspecification.ipynb)
-3. [Principled Bayesian workflow for cognitive models](docs/source/tutorial_notebooks/LCA_Model_Posterior_Estimation.ipynb)
-4. [Posterior estimation for ODEs](docs/source/tutorial_notebooks/Linear_ODE_system.ipynb)
-5. [Posterior estimation for SIR-like models](docs/source/tutorial_notebooks/Covid19_Initial_Posterior_Estimation.ipynb)
-6. [Model comparison for cognitive models](docs/source/tutorial_notebooks/Model_Comparison_MPT.ipynb)
-7. [Hierarchical model comparison for cognitive models](docs/source/tutorial_notebooks/Hierarchical_Model_Comparison_MPT.ipynb)
+1. [Quickstart amortized posterior estimation](examples/Intro_Amortized_Posterior_Estimation.ipynb)
+2. [Tackling strange bimodal distributions](examples/TwoMoons_Bimodal_Posterior.ipynb)
+3. [Detecting model misspecification in posterior inference](examples/Model_Misspecification.ipynb)
+4. [Principled Bayesian workflow for cognitive models](examples/LCA_Model_Posterior_Estimation.ipynb)
+5. [Posterior estimation for ODEs](examples/Linear_ODE_system.ipynb)
+6. [Posterior estimation for SIR-like models](examples/Covid19_Initial_Posterior_Estimation.ipynb)
+7. [Model comparison for cognitive models](examples/Model_Comparison_MPT.ipynb)
+8. [Hierarchical model comparison for cognitive models](examples/Hierarchical_Model_Comparison_MPT.ipynb)
 
 ## Project Documentation
 
@@ -83,7 +84,7 @@ amortized_posterior = bf.amortizers.AmortizedPosterior(inference_net, summary_ne
 Finally, we connect the networks with the generative model via a `Trainer` instance:
 
 ```python
-trainer = bf.trainers.Trainer(amortizer=amortized_posterior, generative_model=generative_model)
+trainer = bf.trainers.Trainer(amortizer=amortized_posterior, generative_model=generative_model, memory=True)
 ```
 
 We are now ready to train an amortized posterior approximator. For instance,
@@ -141,18 +142,17 @@ model parameters as a result of conditioning on the data.
 
 - Radev, S. T., Mertens, U. K., Voss, A., Ardizzone, L., & Köthe, U. (2020).
 BayesFlow: Learning complex stochastic models with invertible neural networks.
-<em>IEEE Transactions on Neural Networks and Learning Systems</em>, available
-for free at: https://arxiv.org/abs/2003.06281.
+<em>IEEE Transactions on Neural Networks and Learning Systems, 33(4)</em>, 1452-1466.
 
 - Radev, S. T., Graw, F., Chen, S., Mutters, N. T., Eichel, V. M., Bärnighausen, T., & Köthe, U. (2021).
-OutbreakFlow: Model-based Bayesian inference of disease outbreak dynamics with invertible neural networks and its application to the COVID-19 pandemics in Germany. <em>PLoS computational biology</em>, 17(10), e1009472.
[…]
