### 2023-07-25 298d4b6
Fixes #90 (axis labeling)
