### 2023-07-27 #5640 @sandeshkatakam
> 👋 @idoby, @sandeshkatakam, @jorgedch, would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

I want to take up this. I had some previous experience with Bayesian Neural networks, which might help me review this better. Let me know the procedure for reviewing this(I just joined as reviewer for JOSS recently)
