### 2023-07-27 #5640 @osorensen
👋 @cranmer, @cthoyt, @LoryPack, given your expertise in the area, would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
