### 2023-07-28 1d7cb0f
fix dict key naming inconsistency ('targets' -> 'parameters') [skip ci]
