### 2023-07-28 d32ef29
Fix train_from_presimulation for model comparison
