### 2023-08-13 ae7f8c4
Add option to aggregate outputs of c2st
