### 2023-08-25 #5702 @osorensen
@sandeshkatakam, please open an issue in the [source repository](https://github.com/stefanradev93/BayesFlow) but include a link to this issue (i.e., `https://github.com/openjournals/joss-reviews/issues/5702`) in the issue that you create. That way we keep track.

For smaller issues you're also welcome to post them right here in the review thread.
