### 2023-08-25 stefanradev93/BayesFlow#96 @sandeshkatakam
Regarding-your-joss-submission
Link to the joss review issue :
[https://github.com/openjournals/joss-reviews/issues/5702](https://github.com/openjournals/joss-reviews/issues/5702)
This issue is regarding your submission of the software paper BayesFlow to JOSS. I was trying to run the tests you mentioned in the tests folder using pytest. I am encountering the error because of the pyproject.toml file(this line in particular --cov BayesFlow). Can you please look into it once and respond.
