### 2023-08-25 stefanradev93/BayesFlow#96 @marvinschmitt
Thank you for opening this issue.

Could you please try to install `pytest-cov` for the coverage metrics and see whether that resolves the issue?

```
pip install pytest-cov
```

Thanks,
Marvin
