### 2023-08-25 stefanradev93/BayesFlow#96 @yannikschaelte
Hi @sandeshkatakam , the prerequisites for development should also be described here: https://bayesflow.org/contributing.html#environment

Specifically, running `pip install -r requirements_dev.txt` should install all necessary dependencies. @marvinschmitt it seems https://github.com/stefanradev93/BayesFlow/blob/Development/requirements_dev.txt could be updated, esp. to only require `>=` minimum versions?
