### 2023-08-25 stefanradev93/BayesFlow#96 @stefanradev93
I have updated `requirements_dev.txt` in `Development` to use at least the recent versions of the listed packages (`>=`) .
