### 2023-08-27 stefanradev93/BayesFlow#96 @sandeshkatakam
Thank you. The tests are running fine now.
