### 2023-08-29 e5e9437
Add bidirectional option to recurrent nets
