### 2023-08-29 0f53c8a
Update changelog to reflect bidirectional change
