### 2023-08-29 17551f5
Reflect name change SequentialNetwork -> SequenceNetwork
