### 2023-08-29 3cee5ae
Update SequenceNetwork to SequentialNetwork
