+++ bayesflow/default_settings.py
@@ -98,7 +98,7 @@ DEFAULT_SETTING_DENSE_PMP = {
 
 DEFAULT_SETTING_AFFINE_COUPLING = MetaDictSetting(
     meta_dict={
-        "dense_args": dict(units=128, activation="relu", kernel_regularizer=tf.keras.regularizers.l2(5e-4)),
+        "dense_args": dict(units=128, activation="relu", kernel_regularizer=tf.keras.regularizers.l2(1e-4)),
         "num_dense": 2,
         "spec_norm": False,
         "mc_dropout": False,
@@ -113,7 +113,7 @@ DEFAULT_SETTING_AFFINE_COUPLING = MetaDictSetting(
 
 DEFAULT_SETTING_SPLINE_COUPLING = MetaDictSetting(
     meta_dict={
-        "dense_args": dict(units=128, activation="relu", kernel_regularizer=tf.keras.regularizers.l2(1e-3)),
+        "dense_args": dict(units=128, activation="relu", kernel_regularizer=tf.keras.regularizers.l2(1e-4)),
         "num_dense": 2,
         "spec_norm": False,
         "mc_dropout": False,
