### 2023-08-30 eb28ca5
Change default of SetTransformer to not use LayerNorm -> better performance found
