### 2023-08-30 4061488
Dropout = 0.05 as default
