### 2023-08-31 eb911c7
Aesthetisize plot_calibration_curves
