### 2023-09-03 stefanradev93/BayesFlow#97 @LoryPack
JOSS review: citations missing
https://github.com/stefanradev93/BayesFlow/blob/45389ad93ae07582ffb1249f407ab09c1b49c6de/JOSS/paper.md?plain=1#L83

I believe the paper should also cite ABCpy here: https://github.com/eth-cscs/abcpy, with paper [here](https://www.jstatsoft.org/article/view/v100i07)

Related to https://github.com/openjournals/joss-reviews/issues/5702
