+++ JOSS/paper.bib
@@ -5,8 +5,19 @@
   JOURNALTITLE = {Osdi},
   NUMBER = {2016},
   PAGES = {265--283},
-  TITLE = {Tensorflow: a system for large-scale machine learning.},
+  TITLE = {{TensorFlow}: A system for large-scale machine learning.},
   VOLUME = {16},
+}
+
+@ARTICLE{dutta2021abcpy,
+ TITLE = {{ABCpy}: A high-performance computing perspective to approximate {B}ayesian computation},
+ VOLUME = {100},
+ DOI = {10.18637/jss.v100.i07},
+ NUMBER = {7},
+ JOURNALTITLE = {Journal of Statistical Software},
+ AUTHOR = {Dutta, Ritabrata and Schoengens, Marcel and Pacchiardi, Lorenzo and Ummadisingu, Avinash and Widmer, Nicole and K\"unzli, Pierre and Onnela, Jukka-Pekka and Mira, Antonietta},
+ DATE = {2021},
+ PAGES = {1–38}
 }
 
 @ARTICLE{bieringer2021measuring,
@@ -16,7 +27,7 @@
   NUMBER = {6},
   PAGES = {126},
   TITLE = {Measuring QCD splittings with invertible networks},
-  doi = {10.21468/SciPostPhys.10.6.126},
+  DOI = {10.21468/SciPostPhys.10.6.126},
   VOLUME = {10},
 }
 
