+++ JOSS/paper.md
@@ -81,7 +81,7 @@ The software is built on top of `TensorFlow` [@abadi2016tensorflow] and thereby 
 
 # Related Software
 When a non-amortized inference procedure does not create a computational bottleneck, approximate Bayesian computation (ABC) might be an appropriate tool. This is the case if a single data set needs to be analyzed, if an infrastructure for parallel computing is readily available, or if repeated re-fits of a model (e.g., cross-validation) are not desired.
-A variety of mature Python packages for ABC exist, such as PyMC [@Salvatier2016], pyABC [@schaelte2022pyabc], or ELFI [@lintusaari2018elfi]. In contrast to these packages, `BayesFlow` focuses on amortized inference, but can also interact with ABC samplers (e.g., use BayesFlow to learn informative summary statistics for an ABC analysis).
+A variety of mature Python packages for ABC exist, such as PyMC [@Salvatier2016], pyABC [@schaelte2022pyabc], ABCpy [@dutta2021abcpy], or ELFI [@lintusaari2018elfi]. In contrast to these packages, `BayesFlow` focuses on amortized inference, but can also interact with ABC samplers (e.g., use BayesFlow to learn informative summary statistics for an ABC analysis).
 
 When it comes to simulation-based inference with neural networks, the `sbi` toolkit enables both likelihood and posterior estimation using different inference algorithms, such as Neural Posterior Estimation [@papamakarios2021normalizing], Sequential Neural Posterior Estimation [@greenberg2019automatic] and Sequential Neural Likelihood Estimation [@papamakarios2019sequential]. `BayesFlow` and `sbi` can be viewed as complementary toolkits, where `sbi` implements a variety of different approximators for standard modeling scenarios, while `BayesFlow` focuses on amortized workflows with user-friendly default settings and optional customization. The `Swyft` library focuses on Bayesian parameter inference in physics and astronomy. `Swyft` uses a specific type of simulation-based neural inference technique, namely, Truncated Marginal Neural Ratio Estimation [@miller2021truncated]. This method improves on standard Markov chain Monte Carlo (MCMC) methods for ABC by learning the likelihood-to-evidence ratio with neural density estimators. Finally, the `Lampe` library provides implementations for a subset of the methods for posterior estimation in the `sbi` library, aiming to expose all components (e.g., network architectures, optimizers) in order to provide a customizable interface for creating neural approximators. All of these libraries are built on top of `PyTorch`.
 
