### 2023-09-03 stefanradev93/BayesFlow#97 @stefanradev93
@LoryPack Thanks for pointing us at the ABCpy library! I have now added the appropriate .bib entry and reference in the text.
