### 2023-09-04 9464595
Fix ndim -> rank bug
