+++ bayesflow/coupling_networks.py
@@ -436,9 +436,9 @@ class SplineCoupling(tf.keras.Model):
         """
 
         shape = tf.shape(parameters)
-        if tf.rank(parameters) == 2:
+        if len(shape) == 2:
             new_shape = (shape[0], self.dim_out, -1)
-        elif tf.rank(parameters) == 3:
+        elif len(shape) == 3:
             new_shape = (shape[0], shape[1], self.dim_out, -1)
         else:
             raise NotImplementedError("Spline flows can currently only operate on 2D and 3D inputs!")
