### 2023-09-04 f46374d
Bugfix ranks for various shapes 3
