+++ bayesflow/helper_networks.py
@@ -112,7 +112,7 @@ class DenseCouplingNet(tf.keras.Model):
 
         # Handle 3D case for a set-flow and repeat condition over
         # the second `time` or `n_observations` axis of `target``
-        if tf.rank(target) == 3 and tf.rank(condition) == 2:
+        if len(tf.shape(target)) == 3 and len(tf.shape(condition)) == 2:
             shape = tf.shape(target)
             condition = tf.expand_dims(condition, 1)
             condition = tf.tile(condition, [1, shape[1], 1])
