### 2023-09-04 f3196e4
Update paper.bib (proper capitalization)
