+++ JOSS/paper.bib
@@ -5,7 +5,7 @@
   JOURNALTITLE = {Osdi},
   NUMBER = {2016},
   PAGES = {265--283},
-  TITLE = {{TensorFlow}: A system for large-scale machine learning.},
+  TITLE = {{TensorFlow}: A system for large-scale machine learning},
   VOLUME = {16},
 }
 
@@ -26,7 +26,7 @@
   JOURNALTITLE = {SciPost Physics},
   NUMBER = {6},
   PAGES = {126},
-  TITLE = {Measuring QCD splittings with invertible networks},
+  TITLE = {Measuring {QCD} splittings with invertible networks},
   DOI = {10.21468/SciPostPhys.10.6.126},
   VOLUME = {10},
 }
@@ -42,10 +42,8 @@
 }
 
 @ARTICLE{burkner_models_2022,
-  ABSTRACT = {Probabilistic (Bayesian) modeling has experienced a surge of applications in almost all quantitative sciences and industrial areas. This development is driven by a combination of several factors, including better probabilistic estimation algorithms, flexible software, increased computing power, and a growing awareness of the benefits of probabilistic learning. However, a principled Bayesian model building workflow is far from complete and many challenges remain. To aid future research and applications of a principled Bayesian workflow, we ask and provide answers for what we perceive as two fundamental questions of Bayesian modeling, namely (a) "What actually is a Bayesian model?" and (b) "What makes a good Bayesian model?". As an answer to the first question, we propose the PAD model taxonomy that defines four basic kinds of Bayesian models, each representing some combination of the assumed joint distribution of all (known or unknown) variables (P), a posterior approximator (A), and training data (D). As an answer to the second question, we propose ten utility dimensions according to which we can evaluate Bayesian models holistically, namely, (1) causal consistency, (2) parameter recoverability, (3) predictive performance, (4) fairness, (5) structural faithfulness, (6) parsimony, (7) interpretability, (8) convergence, (9) estimation speed, and (10) robustness. Further, we propose two example utility decision trees that describe hierarchies and trade-offs between utilities depending on the inferential goals that drive model building and testing.},
   AUTHOR = {Bürkner, Paul-Christian and Scholz, Maximilian and Radev, Stefan T.},
   DATE = {2022},
-  FILE = {arXiv Fulltext PDF:/Users/paul.buerkner/Zotero/storage/RXAMU46Q/B\"{u}rkner et al. - 2022 - Some models are useful, but how do we know which o.pdf:application/pdf;arXiv.org Snapshot:/Users/paul.buerkner/Zotero/storage/QIBQDVLQ/2209.html:text/html},
   JOURNALTITLE = {arXiv preprint},
   KEYWORDS = {Statistics - Methodology},
   NOTE = {arXiv:2209.02439},
@@ -59,7 +57,7 @@
   DATE = {2020},
   JOURNALTITLE = {PeerJ},
   PAGES = {e10316},
-  TITLE = {A Bayesian brain model of adaptive behavior: an application to the Wisconsin Card Sorting Task},
+  TITLE = {A {Bayesian} brain model of adaptive behavior: an application to the Wisconsin Card Sorting Task},
   VOLUME = {8},
 }
 
@@ -76,7 +74,7 @@
[…]
