### 2023-09-04 #5702 @marvinschmitt
Thank you very much for the kind feedback @osorensen. Also thank you @LoryPack and @sandeshkatakam for your reviews. We appreciate it!

@osorensen thanks for catching those issues, I have updated the `.bib` source accordingly in
 https://github.com/stefanradev93/BayesFlow/commit/f3196e4b5140923d4db5f6535acb839828761d8f
