+++ JOSS/paper.bib
@@ -373,3 +373,18 @@
   DOI = {10.1007/s13349-022-00638-5}
 }
 
+@article{schumacher2023,
+  title = {Neural Superstatistics for {Bayesian} Estimation of Dynamic Cognitive Models},
+  author = {Schumacher, Lukas and Bürkner, Paul-Christian and Voss, Andreas and Köthe, Ullrich and Radev, Stefan T},
+  date = {2023-08-23},
+  journaltitle = {Scientific Reports},
+  shortjournal = {Sci Rep},
+  volume = {13},
+  number = {1},
+  pages = {13778},
+  publisher = {{Nature Publishing Group}},
+  issn = {2045-2322},
+  doi = {10.1038/s41598-023-40278-3},
+  issue = {1}
+}
+
