### 2023-09-08 14e3ed1
Update paper.md
add another recently published application of the software
