+++ JOSS/paper.md
@@ -75,7 +75,7 @@ Currently, the software features four key capabilities for enhancing Bayesian wo
 4. **Model misspecification detection:** Ensure that the resulting posteriors are faithful approximations of the otherwise intractable target posterior, even when simulations do not perfectly represent reality [@schmitt2021detecting; @radev2023jana].
 
 
-`BayesFlow` has been used for amortized Bayesian inference in various areas of applied research, such as epidemiology [@radev2021outbreakflow], cognitive modeling [@von2022mental; @wieschen2020jumping; @sokratous2023ask], computational psychiatry [@d2020bayesian], neuroscience [@ghaderi2022general], particle physics [@bieringer2021measuring], agent-based econometrics models [@shiono2021estimation], seismic imaging [@siahkoohi2023reliable], user behavior [@moon2023amortized], structural health monitoring [@zeng2023probabilistic], aerospace [@tsilifis2022inverse] and wind turbine design [@noever2022model], micro-electro-mechanical systems testing [@heringhaus2022towards], and fractional Brownian motion [@verdier2022variational].
+`BayesFlow` has been used for amortized Bayesian inference in various areas of applied research, such as epidemiology [@radev2021outbreakflow], cognitive modeling [@von2022mental; @wieschen2020jumping; @sokratous2023ask; @schumacher2023], computational psychiatry [@d2020bayesian], neuroscience [@ghaderi2022general], particle physics [@bieringer2021measuring], agent-based econometrics models [@shiono2021estimation], seismic imaging [@siahkoohi2023reliable], user behavior [@moon2023amortized], structural health monitoring [@zeng2023probabilistic], aerospace [@tsilifis2022inverse] and wind turbine design [@noever2022model], micro-electro-mechanical systems testing [@heringhaus2022towards], and fractional Brownian motion [@verdier2022variational].
 
 The software is built on top of `TensorFlow` [@abadi2016tensorflow] and thereby enables off-the-shelf support for GPU and TPU acceleration. Furthermore, it can seamlessly interact with TensorFlow Probability [@dillon2017tensorflow] for flexible latent distributions and a variety of joint priors.
 
