+++ JOSS/paper.bib
@@ -384,7 +384,6 @@
   pages = {13778},
   publisher = {{Nature Publishing Group}},
   issn = {2045-2322},
-  doi = {10.1038/s41598-023-40278-3},
-  issue = {1}
+  doi = {10.1038/s41598-023-40278-3}
 }
 
