+++ JOSS/paper.bib
@@ -253,12 +253,13 @@
   VOLUME = {7},
 }
 
-@ARTICLE{schmitt2021detecting,
-  AUTHOR = {Schmitt, Marvin and Bürkner, Paul-Christian and Köthe, Ullrich and Radev, Stefan T},
-  DATE = {2021},
-  JOURNALTITLE = {arXiv preprint},
-  PAGES = {arXiv--2112},
-  TITLE = {Detecting Model Misspecification in Amortized {Bayesian} Inference with Neural Networks},
+
+
+@article{schmitt2021detecting,
+  Author = {Marvin Schmitt and Paul-Christian Bürkner and Ullrich Köthe and Stefan T. Radev},
+  Title = {Detecting Model Misspecification in Amortized {B}ayesian Inference with Neural Networks},
+  year = {2023},
+  journal = {45th German Conference on Pattern Recognition (GCPR)}
 }
 
 @ARTICLE{schmitt2022meta,
