### 2023-09-09 f806091
Update 2d scatterplots to work with recent seaborn
