### 2023-09-09 4e9d475
Fix legend bug in plot_losses without validation loss
