### 2023-09-10 b3c562c
Fix MMD shape determination for  offline autograph mode.
