### 2023-09-10 0709a3c
Docfix and raise ConfigurationError if summary_loss_fun is provided without a summary_net
