### 2023-09-13 4d429ae
set_label names in plot_recovery()
qol change to allow to change label names if wanting something other than "ground truth" / "estimated"
