### 2023-09-14 #5702 @marvinschmitt
@osorensen Here we go, thanks for your patience :-)

- Tagged release: JOSS (v1.1.4)     [[Link]](https://github.com/stefanradev93/BayesFlow/releases/tag/v1.1.4)
- DOI: 10.5281/zenodo.8346393   [[Link](https://doi.org/10.5281/zenodo.8346393)]
