### 2023-09-14 #5702 @osorensen
@openjournals/dev, could you please help with the error message I get when recommending acceptance?
