### 2023-09-14 #5702 @danielskatz
The doi field should be 10.1103/PhysRevE.106.055311 - not just PhysRevE.106.055311
