### 2023-09-15 91b4020
fix doi entry
as noticed in the JOSS review https://github.com/openjournals/joss-reviews/issues/5702
