+++ JOSS/paper.bib
@@ -336,7 +336,7 @@
   PAGES = {055311},
   TITLE = {Variational inference of fractional {Brownian} motion with linear computational complexity},
   VOLUME = {106},
-  DOI = {PhysRevE.106.055311}
+  DOI = {10.1103/PhysRevE.106.055311}
 }
 
 @ARTICLE{von2022mental,
