### 2023-09-15 #5702 @marvinschmitt
@osorensen @danielskatz thank you for tracking down that issue, we appreciate it. I have updated the `doi` field of that bib entry.
