### 2023-09-18 4d415e6
Update readme with better practices [skip ci]
