### 2023-09-18 70cc501
Small changes to doc and **kwargs
